+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mojang's textures queue now caches textures data for 70 seconds to avoid strange `429` errors.
 - Mojang's textures queue now doesn't log timeout errors.
 
+### Fixed
+- Panic when Redis connection is broken.
+- Duplication of Redis connections pool for Mojang's textures queue.
+
 ## [4.2.0] - 2019-05-02
 ### Added
 - `CHANGELOG.md` file.